+++ src/main/java/com/epam/hiltonbooking/servlets/DeleteUserServlet.java
@@ -8,6 +8,7 @@ import jakarta.servlet.annotation.WebServlet;
 import jakarta.servlet.http.HttpServlet;
 import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
+import jakarta.servlet.http.HttpSession;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 
@@ -20,14 +21,24 @@ public class DeleteUserServlet extends HttpServlet {
 
     @Override
     protected void doPost(HttpServletRequest req, HttpServletResponse resp) throws ServletException, IOException {
+        HttpSession session = req.getSession();
         Optional<String> userId = Optional.ofNullable(req.getParameter("user-id"));
 
         UserService userService = ServiceFactory.getInstance().getUserService();
 
+        String infoMessage = null;
         try {
-            if(userId.isPresent()) {
-                userService.deleteUserById(Integer.valueOf(userId.get()));
+            if (userId.isPresent()) {
+                if (userService.hasBookings(Integer.parseInt(userId.get()))) {
+                    infoMessage = "User has booking(s). You cannot delete user.";
+                } else {
+                    infoMessage = "Successfully deleted!";
+                    userService.deleteUserById(Integer.valueOf(userId.get()));
+                }
+                session.setAttribute("infoMessage", infoMessage);
                 resp.sendRedirect(req.getContextPath() + "/users");
+            } else {
+                resp.sendRedirect(req.getContextPath() + "/error");
             }
         } catch (ServiceException e) {
             logger.error("Unable to delete user!");
